Ambigious pow

+++ week1/blocked.cpp
@@ -123,7 +123,7 @@ int bs_scan_search_iter(int q, int* arr, int n, int block)
 int main(int argc, char* argv[])
 {
   int i = 3;
-  int n = pow(d, i) - 1;
+  int n = (int)pow((double)d, i) - 1;
   int* arr = (int*) malloc(n * sizeof(int));
   int* numbers = (int*) malloc(n * sizeof(int));
